refractored code to use firebase.ts
