+++ components/MovieTile.tsx
@@ -1,9 +1,8 @@
 import {Card, CardImg, Button} from "react-bootstrap";
 import Link from "next/link";
-import {setDoc} from "@firebase/firestore";
-import {doc, getDoc, getFirestore} from "firebase/firestore";
 import {UserContext} from "../lib/context";
 import {useContext} from "react";
+import {deleteMovieFromList} from "../lib/firebase";
 
 type propsType = {
     movie: Movie;
@@ -12,32 +11,13 @@ type propsType = {
 
 
 export default function MovieTile(props: propsType) {
-    const db = getFirestore()
-    async function fetchUserInfo() {
-        console.log("read")
-        return getDoc(doc(db,"users", user.uid))
 
-    }
-    async function deleteMovieFromList(id: number, list: string) {
-
-        fetchUserInfo().then((res) => {
-            var listedMovies: ListedMovie[] = res.get("listedMovies")
-
-            listedMovies[listedMovies.findIndex(x => x.movieId == id)].lists = listedMovies[listedMovies.findIndex(x => x.movieId == id)].lists.filter(x => x !== list)
-            let data = {
-                listedMovies: listedMovies
-            }
-            setDoc(doc(db,"users", user.uid), data, { merge:true })
-
-        })
-
-    }
     const {user} = useContext(UserContext);
     // @ts-ignore
     return (
         <div className={"m-2"} data-cy={`MovieTile`}>
             <Card className="bg-light-gray cardAnim" style={{width: "14rem"}}>
-                {props.list && <Button className={"xButton"} onClick={async () =>{await deleteMovieFromList(props.movie.id,props.list!)}}>X</Button>}
+                {props.list && <Button className={"xButton"} onClick={async () =>{await deleteMovieFromList(props.movie.id,props.list!,user.uid)}}>X</Button>}
                 <Link href={"/movies/" + props.movie.id} className={"text-decoration-none text-reset"}>
                     <CardImg src={"https://image.tmdb.org/t/p/w500" + props.movie.poster_path} alt="image of movie" />
                     <Card.Body className={"cardTextPos"}>
